rename crypto_verify to crypto_verify_16
Before this patch, it was arguably not explicit enough that `crypto_verify` is working on buffers of exactly 16 bytes each.

In NaCL this function is exposed as `crypto_verify_16`. If this document promote copying the function, it is perhaps better to have the example match the exposed name of NaCL (rather its actual name in NaCL source which is an implementation detail).